requirements.txt
numpy==1.26.4
pandas==2.1.4
python-dotenv==1.0.1
Flask==3.0.3
SQLAlchemy==2.0.32
pydantic==2.8.2
